impl(server,web,cli): extract @forager/server (tauri port phase 1)
Carve the RPC endpoint, media-file streaming, thumbnail streaming, and the
Forager config schema out of the SvelteKit app into a new framework-agnostic
@forager/server package.

@forager/server
  - api.ts (moved from packages/web/src/lib/api.ts)
  - config.ts (moved from packages/web/src/lib/server/config.ts)
  - handlers/rpc.ts (wraps ts-rpc's SvelteKit adapter with a Request stub
    until ts-rpc publishes a Request adapter; module is internal)
  - handlers/media_file.ts, handlers/thumbnail.ts (pure Request -> Response)
  - ForagerServer class with try_handle_request(request) so embedders can
    mount it ahead of their own routes, plus a standalone start() for the
    upcoming 'forager serve' command.

@forager/web
  - SvelteKit +server.ts routes become thin delegates so dev mode (vite +
    sveltekit hooks) keeps working unchanged.
  - The published web.Server now accepts forager + config directly and
    dispatches /rpc/* and /files/* through ForagerServer, bypassing
    SvelteKit. Static assets and HTML routes still flow through SvelteKit.
  - $lib/api.ts and $lib/server/config.ts become shims re-exporting from
    @forager/server so frontend imports don't need to move yet.

@forager/cli
  - Pass forager + config into web.Server in both 'init' and 'gui' so the
    bypass kicks in.

No frontend behaviour changes. Same URLs, same response shapes.
