+++ deno.json
@@ -1,6 +1,7 @@
 {
   "workspace": [
     "./packages/core",
+    "./packages/server",
     "./packages/web",
     "./packages/web/adapter/lib",
     "./packages/cli"
