+++ packages/web/adapter/lib/deno.json
@@ -12,7 +12,6 @@
     "@andykais/ts-rpc": "jsr:@andykais/ts-rpc@^0.2.2",
     "@std/http/file-server": "jsr:@std/http@0.224.5/file-server",
     "svelte": "npm:svelte@~5.39.8"
-
   },
 
   "exports": {
