+++ packages/web/adapter/lib/mod.ts
@@ -2,6 +2,8 @@ import {serveDir, serveFile, type ServeDirOptions} from '@std/http/file-server';
 import * as path from '@std/path'
 import {Forager} from '@forager/core'
 import {Logger, type LogLevel} from '@forager/core/logger'
+import { ForagerServer } from '@forager/server'
+import { type Config } from '@forager/server/config'
 import { trace } from "@opentelemetry/api"
 
 // these relative imports are generated by the adapter in the build step
@@ -28,6 +30,12 @@ interface ServerOptions {
   // special flag used for development. Disables fetching of assets from jsr.io
   preview?: boolean
 
+  // When provided, /rpc/* and /files/* are dispatched through @forager/server
+  // directly, bypassing SvelteKit. SvelteKit still handles all other (HTML)
+  // routes. When omitted, requests fall through to SvelteKit as before.
+  forager?: Forager
+  config?: Config
+
   kit?: {
     // the key/value pair of environment variables handed to svelte kit server
     env?: Record<string, string>
@@ -43,6 +51,7 @@ class Server {
   #rootDir: string
   #kitServerInitialized!: Promise<void>
   #logger: Logger
+  #forager_server?: ForagerServer
   #routes: {
     favicon: URLPattern
     immutable_asset: URLPattern
@@ -75,6 +84,13 @@ class Server {
     this.#serve_dir_options = {
       fsRoot: this.#rootDir,
       quiet: this.#options.logger?.level !== 'DEBUG',
+    }
+    if (options.forager && options.config) {
+      this.#forager_server = new ForagerServer({
+        forager: options.forager,
+        config: options.config,
+        logger: options.logger,
+      })
     }
   }
 
@@ -238,6 +254,13 @@ class Server {
       }
     }
 
+    // When the embedder gave us a Forager instance, dispatch /rpc/* and
+    // /files/* through @forager/server directly, bypassing SvelteKit.
+    if (this.#forager_server) {
+      const forager_response = await this.#forager_server.try_handle_request(request)
+      if (forager_response) return forager_response
+    }
+
     this.#logger.debug(`respond with @sveltekit server`)
     // Pass to the SvelteKit server
     await this.#kitServerInitialized
@@ -269,6 +292,8 @@ if (import.meta.main) {
     logger: config.web.logger,
     asset_folder: path.join(Deno.cwd(), 'static_assets'),
     preview: true,
+    forager,
+    config,
     kit: {
       env,
     }
