+++ packages/web/src/lib/api.ts
@@ -1,41 +1,4 @@
-import { type Forager } from '@forager/core'
-import * as rpc from '@andykais/ts-rpc/adapters/sveltekit.ts'
-import { type Config } from '$lib/server/config.ts'
-
-
-interface Context {
-  forager: Forager
-  config: Config
-}
-
-class ForagerTagApi extends rpc.ApiController<Context> {
-  search = this.context.forager.tag.search
-  get = this.context.forager.tag.get
-  update = this.context.forager.tag.update
-  alias_create = this.context.forager.tag.alias_create
-  alias_delete = this.context.forager.tag.alias_delete
-  parent_create = this.context.forager.tag.parent_create
-  parent_delete = this.context.forager.tag.parent_delete
-}
-
-class ForagerApi extends rpc.ApiController<Context> {
-  media = this.context.forager.media
-  series = this.context.forager.series
-  tag = this.module(ForagerTagApi)
-  views = this.context.forager.views
-
-}
-
-export class Api extends rpc.ApiController<Context> {
-  forager = this.module(ForagerApi)
-
-  config(): Config {
-    return this.context.config
-  }
-
-  server_time(): Date {
-    return new Date()
-  }
-}
-
-export type ApiSpec = rpc.InferSpec<typeof Api>
+// `@forager/server` owns the RPC `Api` definition as of Phase 1 of the
+// Tauri port. This file remains as a re-export so existing `$lib/api.ts`
+// imports (e.g. `BaseController`'s `rpc.create<ApiSpec>(...)`) keep working.
+export { Api, type ApiSpec, type ApiContext } from '@forager/server/api'
